+++ v4l2uvc.c
@@ -346,7 +346,7 @@ int uvcGrab(struct vdIn *vd)
 
     static int count_t = 0;
 	static int count_last = 0;
-    static unsigned long time_begin, time_current, time_last = 0;
+    static long long time_begin, time_current, time_last = 0;
 	int frame_interval = ( 1000-100 ) / threadcfg.framerate;
 	int time1;
     if (count_t == 0)
